requirements.txt
curl-cffi>=0.5.10
python-telegram-bot>=20.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
psycopg2-binary>=2.9.9
